Implement 12bit unsigned address (#266)
* remove explicit bash terminal setting

* Begin implementing u12 type

* Add wrapping arithmetic to u12

* Pretify the output of examples

* Incorporate u12 into opcodes

+++ src/cpu/chip8/operations/mod.rs
@@ -27,7 +27,7 @@ impl Generate<Chip8, Vec<Microcode>> for opcodes::Jp {
     fn generate(self, _: &Chip8) -> Vec<Microcode> {
         vec![Microcode::Write16bitRegister(Write16bitRegister::new(
             register::WordRegisters::ProgramCounter,
-            self.addr().wrapping_sub(2),
+            u16::from(self.addr()).wrapping_sub(2),
         ))]
     }
 }
